+++ client/package.json
@@ -4,18 +4,19 @@
   "private": true,
   "dependencies": {
     "@date-io/date-fns": "^1.3.13",
-    "@material-ui/pickers": "^3.2.10",
-    "@toast-ui/react-editor": "^2.3.1",
-    "date-fns": "^2.16.0",
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.56",
+    "@material-ui/pickers": "^3.2.10",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
+    "@toast-ui/react-editor": "^2.3.1",
     "axios": "^0.20.0",
     "connected-react-router": "^6.8.0",
+    "date-fns": "^2.16.0",
     "formik": "^2.1.5",
+    "google-maps-react": "^2.0.6",
     "node-sass": "^4.14.1",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
@@ -23,11 +24,10 @@
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
-    "tui-editor": "^1.4.10",
-    "use-places-autocomplete": "^1.3.10",
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
     "redux-saga": "^1.1.3",
+    "tui-editor": "^1.4.10",
     "yup": "^0.29.3"
   },
   "scripts": {
