+++ client/public/index.html
@@ -29,7 +29,6 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyB_BJhQ4nBvi7cPxi8DRGJepYp4MbdtRcQ&libraries=places"></script>
     <div id="modal-root"></div>
     <!--
       This HTML file is a template.
