+++ client/src/components/molecule/createEvent/maps.scss
@@ -1,4 +1,8 @@
 .create-event-form {
+  .map-wrap {
+    position: relative;
+    margin-bottom: 400px;
+  }
   #map {
     height: 400px;
     margin-top: 30px;
@@ -13,6 +17,7 @@
   }
   #infowindow-content .title {
     font-weight: bold;
+    margin-left: 4px;
   }
 
   #infowindow-content {
@@ -74,5 +79,9 @@
   }
   #place-name {
     font-size: 1.7rem;
+  }
+  #place-address {
+    display: block;
+    padding-top: 5px;
   }
 }
