+++ client/src/components/template/createEvent/EventForm.jsx
@@ -1,8 +1,9 @@
-import React, { useRef, useState } from 'react';
+import React, { useRef, useState, useEffect } from 'react';
+import { GoogleApiWrapper } from 'google-maps-react';
+import { useDispatch, useSelector } from 'react-redux';
 import EventDislose from '../../molecule/createEvent/EventDisclose';
 import EventTitle from '../../molecule/createEvent/EventTitle';
 import EventOrganizer from '../../molecule/createEvent/EventOrganizer';
-import EventApplyLink from '../../molecule/createEvent/EventApplyLink';
 import EventOnlineCheck from '../../molecule/createEvent/EventOnlineCheck';
 import EventAddress from '../../molecule/createEvent/EventAddress';
 import EventAddressDetail from '../../molecule/createEvent/EventAddressDetail';
@@ -16,7 +17,91 @@ import EventDate from '../../molecule/createEvent/EventDate';
 
 import axios from 'axios';
 
-function EventForm() {
+const EventForm = () => {
+  const [placeState, setPlaceState] = useState({
+    icon: '',
+    name: '',
+    address: '',
+    lat: 0,
+    lng: 0,
+  });
+  useEffect(() => {
+    const map = new window.google.maps.Map(document.getElementById('map'), {
+      center: { lat: 37.5452446, lng: 127.0570452 },
+      zoom: 17,
+    });
+    const card = document.getElementById('pac-card');
+    const input = document.getElementById('pac-input');
+    map.controls[window.google.maps.ControlPosition.TOP_RIGHT].push(card);
+    const autocomplete = new window.google.maps.places.Autocomplete(input);
+    autocomplete.bindTo('bounds', map);
+    autocomplete.setFields(['address_components', 'geometry', 'icon', 'name']);
+    const infowindow = new window.google.maps.InfoWindow();
+    const infowindowContent = document.getElementById('infowindow-content');
+    infowindow.setContent(infowindowContent);
+    const marker = new window.google.maps.Marker({
+      map,
+      anchorPoint: new window.google.maps.Point(0, -29),
+      position: map.center,
+      zoom: 13,
+    });
+
+    infowindowContent.children['place-icon'].src =
+      'https://maps.gstatic.com/mapfiles/place_api/icons/generic_business-71.png';
+    infowindowContent.children['place-name'].textContent = '제강빌딩';
+    infowindowContent.children['place-address'].textContent =
+      '２８９−１０ 성수2가3동 성동구';
+    infowindow.open(map, marker);
+    autocomplete.addListener();
+    autocomplete.addListener('place_changed', () => {
+      infowindow.close();
+      marker.setVisible(false);
+      const place = autocomplete.getPlace();
+
+      if (!place.geometry) {
+        window.alert("No details available for input: '" + place.name + "'");
+        return;
+      }
+
+      if (place.geometry.viewport) {
+        map.fitBounds(place.geometry.viewport);
+      } else {
+        map.setCenter(place.geometry.location);
+        map.setZoom(17);
+      }
+      marker.setPosition(place.geometry.location);
+      marker.setVisible(true);
+      let address = '';
+
+      if (place.address_components) {
+        address = [
+          (place.address_components[0] &&
+            place.address_components[0].short_name) ||
+            '',
+          (place.address_components[1] &&
+            place.address_components[1].short_name) ||
+            '',
+          (place.address_components[2] &&
+            place.address_components[2].short_name) ||
+            '',
+        ].join(' ');
+      }
+      console.dir(place);
+      const placeAddress = {
+        icon: place.icon,
+        name: place.name,
+        address: address,
+        lat: place.geometry.location.lat(),
+        lng: place.geometry.location.lng(),
+      };
+      setPlaceState(placeAddress);
+      infowindowContent.children['place-icon'].src = place.icon;
+      infowindowContent.children['place-name'].textContent = place.name;
+      infowindowContent.children['place-address'].textContent = address;
+      infowindow.open(map, marker);
+    });
+  }, []);
+  const user = useSelector((state) => state.auth.user);
   const [onlineCheck, setOnlineCheck] = useState(false);
 
   const $form = useRef(null);
@@ -24,7 +109,6 @@ function EventForm() {
   const $eventTitle = useRef(null);
   const $phone = useRef(null);
   const $mail = useRef(null);
-  const $forLink = useRef(null);
   const $startDate = useRef(null);
   const $endDate = useRef(null);
   const $isOnline = useRef(null);
@@ -34,21 +118,16 @@ function EventForm() {
   const $onlinePlatform = useRef(null);
   const $price = useRef(null);
   const $maxPerson = useRef(null);
-  // const $onlineWay = useRef(null);
   const $thumbnail = useRef(null);
   const $toast = useRef(null);
 
   function submit(e) {
     e.preventDefault();
-    // const node = document.querySelector('.tui-editor-contents');
-    // console.log(node);
-    // $toast.current;
     const publicRef = {
       curIsOpen: $isOpen.current.checked,
       curEventTitle: $eventTitle.current.value,
       curTel: $phone.current.value,
       curMail: $mail.current.value,
-      curForLink: $forLink.current.value,
       curStartDate: $startDate.current,
       curEndDate: $endDate.current,
       curIsOnline: $isOnline.current.checked,
@@ -86,7 +165,7 @@ function EventForm() {
       isOpen: publicRef.curIsOpen,
       title: publicRef.curEventTitle,
       host: {
-        id: '5f4766ee0c1c3b64fcebbd2b',
+        id: user._id,
         email: publicRef.curMail,
         phone_number: publicRef.curTel,
       },
@@ -106,7 +185,7 @@ function EventForm() {
       online_platform: onlineRef.curPlatform,
       location: {
         name: offlineRef.curAddress,
-        details: offlineRef.curAddressDetail,
+        details: placeState,
         info: offlineRef.curAddressDetailPlus,
       },
       price: publicRef.curPrice, // 입장료
@@ -116,9 +195,8 @@ function EventForm() {
       enlisted_users_id: [], // 해당 이벤트 참여신청을 한 유저들의 배열
     };
     console.log(payload);
-    console.log(startDateValue);
 
-    axios.post('/api/v1/events', payload);
+    // axios.post('/api/v1/events', payload);
   }
   function openToggle(e) {
     e.target.parentNode.classList.toggle('active');
@@ -139,7 +217,6 @@ function EventForm() {
         <EventDislose toggle={openToggle} Ref={$isOpen} />
         <EventTitle Ref={$eventTitle} />
         <EventOrganizer telRef={$phone} mailRef={$mail} />
-        <EventApplyLink Ref={$forLink} />
         <EventDate startDateRef={$startDate} endDateRef={$endDate} />
         <EventOnlineCheck toggle={onlineToggle} Ref={$isOnline} />
 
@@ -170,6 +247,9 @@ function EventForm() {
       </form>
     </div>
   );
-}
+};
 
-export default EventForm;
+export default GoogleApiWrapper({
+  apiKey: 'AIzaSyB_BJhQ4nBvi7cPxi8DRGJepYp4MbdtRcQ',
+  language: 'Korean',
+})(EventForm);
