+++ client/src/pages/createevent.scss
@@ -10,7 +10,7 @@
     padding: 0;
     box-sizing: border-box;
   }
-  .w1300-container {
+  .w1440-container {
     width: 1440px;
     margin: 0 auto;
     padding: 0 30px;
